Mellomlagre ved fortsettSenere
